requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
